requirements.txt
aiosignal==1.4.0
asyncio==4.0.0
asyncpg==0.30.0
attrs==25.3.0
frozenlist==1.7.0
greenlet==3.2.4
idna==3.10
more-itertools==10.7.0
multidict==6.6.4
propcache==0.3.2
python-dotenv==1.1.1
SQLAlchemy==2.0.43
typing_extensions==4.14.1
yarl==1.20.1